Cleaned up the app.js
