+++ api/app.js
@@ -4,10 +4,9 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
-
-
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
+var picsRouter = require('./routes/picsController');
 
 var app = express();
 
@@ -23,6 +22,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
+app.use('/user/:userid', picsRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
